pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["Vitalii Lutsyk <vitaliilutsyk9895@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
openai = "^1.58.1"
uvicorn = "^0.34.0"
requests = "^2.32.3"
types-requests = "^2.32.0.20241016"
redis = "^5.2.1"
hiredis = "^3.1.0"
mypy = "^1.14.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.1"
pytest-redis = "^3.1.3"
pytest-mock-resources = "^2.12.1"
pytest-cov = "^6.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = ["untyped_package.*"]
follow_untyped_imports = true
